+++ RELEASES.md
@@ -1,5 +1,12 @@
 # Release Notes
 
+## [Unreleased]
+
+### Fixed
+
+* zcu102: save errno when cleaning up after init failure
+
+
 ## [v0.4.0] - 2021-05-12
 
 ### Added
@@ -153,6 +160,7 @@
 * Initial public release
 
 
+[Unreleased]: https://github.com/energymon/energymon/compare/v0.4.0...HEAD
 [v0.4.0]: https://github.com/energymon/energymon/compare/v0.3.1...v0.4.0
 [v0.3.1]: https://github.com/energymon/energymon/compare/v0.3.0...v0.3.1
 [v0.3.0]: https://github.com/energymon/energymon/compare/v0.2.4...v0.3.0